ad fs storage engine to testFixtures
