+++ filesystem/build.gradle
@@ -7,6 +7,7 @@ plugins {
     id 'java-library'
     id "io.freefair.lombok"
     id 'jacoco'
+    id 'java-test-fixtures'
 }
 
 ext {
@@ -19,9 +20,9 @@ configurations.all {
 }
 
 dependencies {
-    api project(':core')
+    implementation project(':core')
     // required by hadoop filesystem https://hadoop.apache.org/docs/stable/hadoop-project-dist/hadoop-common/filesystem/index.html.
-    api("org.apache.hadoop:hadoop-common:${hadoop}") {
+    implementation("org.apache.hadoop:hadoop-common:${hadoop}") {
         exclude group: 'org.apache.zookeeper'
         exclude group: 'org.eclipse.jetty'
         exclude group: 'com.sun.jersey'
@@ -67,6 +68,24 @@ dependencies {
     testImplementation group: 'org.hamcrest', name: 'hamcrest-library', version: '2.1'
     testImplementation group: 'org.mockito', name: 'mockito-core', version: '3.12.4'
     testImplementation group: 'org.mockito', name: 'mockito-junit-jupiter', version: '3.12.4'
+    testFixturesImplementation(project(":core"))
+    testFixturesImplementation("org.apache.hadoop:hadoop-common:${hadoop}") {
+        exclude group: 'org.apache.zookeeper'
+        exclude group: 'org.eclipse.jetty'
+        exclude group: 'com.sun.jersey'
+        exclude group: 'javax.servlet.jsp'
+        exclude group: 'javax.servlet'
+        exclude group: 'org.apache.kerby'
+        exclude group: 'org.apache.curator'
+        exclude group: 'com.google.protobuf', module: 'protobuf-java'
+        exclude group: 'org.apache.avro', module: 'avro'
+        exclude group: 'com.nimbusds', module: 'nimbus-jose-jwt'
+        // enforce version.
+        exclude group: 'com.fasterxml.woodstox', module: 'woodstox-core'
+        exclude group: 'commons-io', module: 'commons-io'
+        exclude group: 'ch.qos.reload4j', module: 'reload4j'
+        exclude group: 'org.apache.httpcomponents', module: 'httpcore'
+    }
 }
 
 test {
