+++ integ-test/build.gradle
@@ -93,6 +93,8 @@ dependencies {
     testImplementation group: 'com.h2database', name: 'h2', version: '2.1.210'
     testImplementation group: 'org.xerial', name: 'sqlite-jdbc', version: '3.28.0'
     testImplementation group: 'com.google.code.gson', name: 'gson', version: '2.8.9'
+    testImplementation(testFixtures(project(":core")))
+    testImplementation(testFixtures(project(":filesystem")))
 }
 
 dependencyLicenses.enabled = false
